Moved signature buffer generation code from funder's types.rs to proto.
