+++ components/proto/src/funder/signature_buff.rs
@@ -1,9 +1,11 @@
 use byteorder::{BigEndian, WriteBytesExt};
-use crypto::hash;
+use crypto::hash::{self, sha_512_256, HashResult};
 use crypto::identity::{verify_signature, PublicKey};
 
+use utils::int_convert::usize_to_u64;
+
 use super::messages::{ResponseSendFunds, FailureSendFunds, 
-    SendFundsReceipt, PendingFriendRequest};
+    SendFundsReceipt, PendingFriendRequest, FriendMoveToken};
 
 pub const FUND_SUCCESS_PREFIX: &[u8] = b"FUND_SUCCESS";
 pub const FUND_FAILURE_PREFIX: &[u8] = b"FUND_FAILURE";
@@ -100,5 +102,41 @@ pub fn verify_receipt(receipt: &SendFundsReceipt,
     verify_signature(&data, public_key, &receipt.signature)
 }
 
+
+// Prefix used for chain hashing of token channel funds.
+// NEXT is used for hashing for the next move token funds.
+const TOKEN_NEXT: &[u8] = b"NEXT";
+
+/// Combine all operations into one hash value.
+pub fn operations_hash(friend_move_token: &FriendMoveToken) -> HashResult {
+    let mut operations_data = Vec::new();
+    operations_data.write_u64::<BigEndian>(
+        usize_to_u64(friend_move_token.operations.len()).unwrap()).unwrap();
+    for op in &friend_move_token.operations {
+        operations_data.extend_from_slice(&op.to_bytes());
+    }
+    sha_512_256(&operations_data)
+}
+
+pub fn friend_move_token_signature_buff(friend_move_token: &FriendMoveToken) -> Vec<u8> {
+    let mut sig_buffer = Vec::new();
+    sig_buffer.extend_from_slice(&sha_512_256(TOKEN_NEXT));
+    sig_buffer.extend_from_slice(&operations_hash(friend_move_token));
+    sig_buffer.extend_from_slice(&friend_move_token.old_token);
+    sig_buffer.write_u64::<BigEndian>(friend_move_token.inconsistency_counter).unwrap();
+    sig_buffer.write_u128::<BigEndian>(friend_move_token.move_token_counter).unwrap();
+    sig_buffer.write_i128::<BigEndian>(friend_move_token.balance).unwrap();
+    sig_buffer.write_u128::<BigEndian>(friend_move_token.local_pending_debt).unwrap();
+    sig_buffer.write_u128::<BigEndian>(friend_move_token.remote_pending_debt).unwrap();
+    sig_buffer.extend_from_slice(&friend_move_token.rand_nonce);
+
+    sig_buffer
+}
+
+/// Verify that new_token is a valid signature over the rest of the fields.
+pub fn verify_friend_move_token(friend_move_token: &FriendMoveToken, public_key: &PublicKey) -> bool {
+    let sig_buffer = friend_move_token_signature_buff(friend_move_token);
+    verify_signature(&sig_buffer, public_key, &friend_move_token.new_token)
+}
 
 // TODO: How to test this?
